Update consul discovery mechanism to get list of nodes from couchbase
List of nodes can be updated on consul side but not on the couchbase cluster.
Due to that if we recreate the monitor from the list of nodes in consul we can still have a monitor querying a node that will be removed.
Then when it is really rermoved from couchbase we are not recreating the monitor which leads to some prometheus metrics from this removed node to
be still here.
With that change we will get list of nodes from consul and then query couchbase from thise nde list to get the list of nodes frm couchbase point of view